backend/requirement.txt
# pip install -r requirement.txt 

#install packages
api==0.0.7
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
nose==1.3.7
pillow==11.1.0
PyJWT==1.7.1
python-dotenv==1.0.1
requests==2.32.3
pillow==11.1.0
requirement.txt==0.1.0
sqlparse==0.5.3
urllib3==2.3.0
